Backend/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi[standard]>=0.110.0
uvicorn>=0.25.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
greenlet>=3.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
slowapi>=0.1.9
redis>=5.0.0